+++ bls12_381/src/lib.rs
@@ -4,7 +4,7 @@
 //! curve construction.
 //!
 //! * **This implementation has not been reviewed or audited. Use at your own risk.**
-//! * This implementation targets Rust `1.36` or later.
+//! * This implementation targets Rust `1.88` or later.
 //! * This implementation does not require the Rust standard library.
 //! * All operations are constant time unless explicitly noted.